13.8 Now, lastly, as you're building your app, you're probably going to end up with a lot of different relation managers on the same record and fill in the blank.
allows you to do that really well. So if we go back to our speaker resource, now, I'm just going to add a duplicate of this relation manager, just so that you can see it. But when I add that here, and I go back to the browser here, now, it's not actually going to work because it's a duplicate. But when you have different models, you're going to be able to switch back and forth, you can even customize the icon and how these look. So when you have multiple relationships for a single model, you can add your relation managers here. And you can interact with not just the model, but all its relationships in one place, it's super powerful.